moved level up func into own class
